requirements.txt
absl-py==2.1.0
aif360==0.6.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
autograd==1.6.2
autograd-gamma==0.5.0
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blis==0.7.11
Brotli==1.1.0
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
cloudpickle==3.0.0
colorlog==6.8.2
comm==0.2.2
confection==0.1.5
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
dask==2023.5.0
debugpy==1.8.1
decaf-synthetic-data==0.1.6
decorator==5.1.1
defusedxml==0.7.1
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.2.1
executing==2.0.1
fairlearn==0.10.0
fairx @ file:///home/fahim/code/FairX
fastai==2.7.15
fastcore==1.5.44
fastdownload==0.0.7
fastjsonschema==2.19.1
fastprogress==1.0.3
feather-format==0.4.1
fflows==0.0.3
filelock==3.14.0
fonttools==4.53.0
formulaic==1.0.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.0
future==1.0.0
geomloss==0.2.6
google-auth==2.30.0
google-auth-oauthlib==1.0.0
graphlib_backport==1.1.0
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imbalanced-learn==0.12.3
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflate64==1.0.0
iniconfig==2.0.0
interface-meta==1.3.0
ipykernel==6.29.4
ipython==8.12.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
keopscore==2.2.3
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lifelines==0.27.8
lightning-utilities==0.11.2
llvmlite==0.41.1
locket==1.0.0
loguru==0.7.2
Mako==1.3.5
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
monai==1.3.1
mpmath==1.3.0
multidict==6.0.5
multivolumefile==0.2.3
murmurhash==1.0.10
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.8
nflows==0.14
notebook==7.2.1
notebook_shim==0.2.4
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opacus==1.4.1
opt-einsum==3.3.0
optuna==3.6.1
overrides==7.7.0
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.1
patsy==0.5.6
pexpect==4.9.0
pgmpy==0.1.25
pickleshare==0.7.5
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
pluggy==1.5.0
preshed==3.0.9
prometheus_client==0.20.0
prompt_toolkit==3.0.46
protobuf==5.27.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.21.0
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybcj==1.0.2
pybind11==2.12.0
pycox==0.2.3
pycparser==2.22
pycryptodomex==3.20.0
pydantic==1.10.15
pydantic_core==2.18.4
Pygments==2.18.0
pykeops==2.2.3
pyparsing==3.1.2
pyppmd==1.1.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytorch-lightning==1.9.5
pytorch-ranger==0.1.1
pyts==0.13.0
pytz==2024.1
pyunpack==0.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==26.0.3
pyzstd==0.16.0
redis==5.0.5
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scikit-base==0.7.8
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
Send2Trash==1.8.3
shap==0.44.1
shellingham==1.5.4
six==1.16.0
sktime==0.29.1
slicer==0.0.7
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
stack-data==0.6.3
statsmodels==0.14.1
stumpy==1.12.0
sympy==1.12.1
synthcity==0.2.4
tensorboard==2.14.0
tensorboard-data-server==0.7.2
terminado==0.18.1
texttable==1.7.0
thinc==8.2.4
threadpoolctl==3.5.0
tinycss2==1.3.0
tomli==2.0.1
toolz==0.12.1
torch==2.3.1
torch-optimizer==0.3.0
torchmetrics==1.4.0.post0
torchtext==0.18.0
torchtuples==0.2.2
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
triton==2.3.1
tsai==0.2.17
tsfresh==0.20.2
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.1
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.16.0
xgboost==2.0.3
xgbse==0.2.3
yarl==1.9.4
zipp==3.19.2

